+++ ft_puthexa_upper.c
@@ -6,14 +6,14 @@
 /*   By: oairola <oairola@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 14:08:33 by olwar             #+#    #+#             */
-/*   Updated: 2022/03/04 09:27:32 by oairola          ###   ########.fr       */
+/*   Updated: 2022/03/04 14:24:19 by oairola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include <stdio.h>
 
-static int	ft_puthex_upper_neg(unsigned long int n, int len)
+static int	ft_puthex_upper_neg(unsigned long long n, int len)
 {
 /* 	int	len;
 	
@@ -36,7 +36,7 @@ static int	ft_puthex_upper_neg(unsigned long int n, int len)
 	return (len);
 }
 
-static int	ft_puthex_upper(long int n, int len)
+static int	ft_puthex_upper(unsigned long long int n, int len)
 {
 	unsigned	num;
 /* 	int			len;
@@ -86,7 +86,7 @@ int	ft_puthexa_upper(va_list args, t_node *head, char *format_part)
 		len += ft_puthex_upper(va_arg(args, unsigned long int), 0);
 		/* minimum_number_of_ints (is smaller then just add 0s) */
 	else 
-		len += ft_puthex_upper(va_arg(args, long int), 0);
+		len += ft_puthex_upper(va_arg(args, long long int), 0);
 	if (ft_justify(args2, head, format_part))
 		len += ft_putwidth(args2, head, format_part);
 	return (len);
